Attempted new method of aiming at AprilTags
I think this method would work, however, because the robot does not know its absolute position on the field it cannot properly calculate the angle to turn to. This would most likely be fixable if the initial position of the robot is known, either by hardcoding a position or using vision to estimate one.